module: Modularize netlink into its own files
