+++ src/module/linux/linux.h
@@ -1,21 +1,15 @@
 #pragma once
 // Linux specific definitions
 #include "syscalls.h"
-
-int krf_netlink_broadcast(char *buf, size_t buflen);
-#define KRF_NETLINK_BUF_SIZE 48
-#define KRF_LINK_LOG(...)                                                                          \
-  ({                                                                                               \
-    char buf[KRF_NETLINK_BUF_SIZE]; /* max message size */                                         \
-    snprintf(buf, KRF_NETLINK_BUF_SIZE, __VA_ARGS__);                                              \
-    krf_netlink_broadcast(buf, strlen(buf) + 1);                                                   \
-  })
+#include "netlink.h"
 
 #define KRF_SAFE_WRITE(x) KRF_CR0_WRITE_UNLOCK(x)
 #define KRF_LOG(...)                                                                               \
   ({                                                                                               \
+    char buf[KRF_NETLINK_BUF_SIZE]; /* max message size */                                         \
     printk(KERN_INFO __VA_ARGS__);                                                                 \
-    KRF_LINK_LOG(__VA_ARGS__);                                                                     \
+    snprintf(buf, KRF_NETLINK_BUF_SIZE, __VA_ARGS__);                                              \
+    krf_netlink_broadcast(buf, strlen(buf) + 1);                                                   \
   })
 #define KRF_SYSCALL_TABLE sys_call_table
 #define KRF_TARGETING_PARMS current
